cmd/res/example/configuration.toml: rm AutoEvents
The default example configuration.toml should "just work" as much as possible,
and since the AutoEvents configuration requires a peer MQTT application which is
not provided with the device service, remove AutoEvents from the default one.

Also add a new, differently named configuration.toml which contains the
AutoEvents spec for the device.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>